+++ vite.config.mts
@@ -5,9 +5,9 @@ export default defineConfig(({ mode }) => ({
   build: {
     outDir: "dist/bundled/",
     lib: {
-      entry: [path.resolve(__dirname, "src/bundles/edxAiChatDrawer.tsx")],
-      name: "edxAiChatDrawer",
-      fileName: (format) => `edxAiChatDrawer.${format}.js`,
+      entry: [path.resolve(__dirname, "src/bundles/remoteAiChatDrawer.tsx")],
+      name: "remoteAiChatDrawer",
+      fileName: (format) => `remoteAiChatDrawer.${format}.js`,
     },
   },
   define: {
